# Add class Stopwatch