--- a/OpenSim/Common/Test/testC3DFileAdapter.cpp
+++ b/OpenSim/Common/Test/testC3DFileAdapter.cpp
@@ -78,9 +78,9 @@ void test(const std::string filename) {
     using namespace std;
 
     // The walking C3D files included in this test should not take more
-    // than 40ms on most hardware. We make the max time 100ms to account
+    // than 40ms on most hardware. We make the max time 200ms to account
     // for potentially slower CI machines.
-    const long long MaximumLoadTimeInNs = SimTK::secToNs(0.100);
+    const long long MaximumLoadTimeInNs = SimTK::secToNs(0.200);
     
     Stopwatch watch;
     C3DFileAdapter c3dFileAdapter{};