--- a/OpenSim/Common/osimCommon.h
+++ b/OpenSim/Common/osimCommon.h
@@ -32,6 +32,8 @@
 #include "GCVSpline.h"
 #include "IO.h"
 
+#include "Stopwatch.h"
+
 #include "Scale.h"
 #include "SimmSpline.h"
 #include "Constant.h"