Fixes part of issue #2559 

### Brief summary of changes

- This PR adds a class Stopwatch that provides a very easy way to record and report the duration of code.

### Testing I've completed

- I employed the class in various locations and ensured the output was as expected.

### CHANGELOG.md (choose one)

- updated...

The Doxygen for this PR can be viewed at http://myosin.sourceforge.net/?C=N;O=D; click the folder whose name is this PR's number.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/opensim-org/opensim-core/2578)
<!-- Reviewable:end -->