@chrisdembia 2019-10-28
Yes, the idea is to replace all those calls. I should double-check locally about how the reported times differ.