@chrisdembia 2019-10-28
Would you like me to do the removal of `std::clock`, etc. in this PR? 

In some cases, it doesn't make sense to use Stopwatch, because wall time is also reported.