@chrisdembia 2019-10-30
Okay great. Timings are comparable on testC3DFileAdapter.

master:
```
	C3DFileAdapter 'walking5.c3d' loaded in 20.288ms
	Wrote 'walking5_markers.trc' in 6.859ms
	Wrote'walking5_grfs.sto' in 1.662ms
	Read'walking5_markers.trc' and its standard in 7ms
```

this PR:
```
	C3DFileAdapter 'walking5.c3d' loaded in 27 millisecond(s)
	Wrote 'walking5_markers.trc' in 6 millisecond(s)
	Wrote'walking5_grfs.sto' in 1 millisecond(s)
	Read'walking5_markers.trc' and its standard in 8 millisecond(s)
```